Revert "build: Moving more downstream code into the listener extension (#24495)"
This reverts commit d78e0c8ce60d8a68ba7281a1252dd172847a7627.
