+++ contrib/network/connection_balance/dlb/source/connection_balancer_impl.h
@@ -10,7 +10,7 @@
 #include "source/common/api/os_sys_calls_impl.h"
 #include "source/common/network/connection_balancer_impl.h"
 #include "source/common/protobuf/protobuf.h"
-#include "source/extensions/listener_managers/listener_manager/active_tcp_listener.h"
+#include "source/server/active_tcp_listener.h"
 
 #include "contrib/envoy/extensions/network/connection_balance/dlb/v3alpha/dlb.pb.h"
 #include "contrib/envoy/extensions/network/connection_balance/dlb/v3alpha/dlb.pb.validate.h"
