+++ source/extensions/bootstrap/internal_listener/active_internal_listener.cc
@@ -6,7 +6,7 @@
 #include "source/common/network/address_impl.h"
 #include "source/common/stats/timespan_impl.h"
 #include "source/extensions/io_socket/user_space/io_handle.h"
-#include "source/extensions/listener_managers/listener_manager/active_stream_listener_base.h"
+#include "source/server/active_stream_listener_base.h"
 
 namespace Envoy {
 namespace Extensions {
