+++ source/extensions/listener_managers/listener_manager/BUILD
@@ -28,7 +28,6 @@ envoy_cc_extension(
         "//test:__subpackages__",
     ],
     deps = [
-        ":connection_handler_lib",
         "//source/server:listener_manager_factory_lib",
         "//source/server:api_listener_lib",
         ":active_raw_udp_listener_config",
@@ -92,8 +91,8 @@ envoy_cc_library(
         "//test:__subpackages__",
     ],
     deps = [
-        ":connection_handler_lib",
         "//envoy/registry",
+        "//source/server:connection_handler_lib",
     ],
 )
 
@@ -152,141 +151,5 @@ envoy_cc_library(
         "@envoy_api//envoy/config/listener/v3:pkg_cc_proto",
         "@envoy_api//envoy/config/route/v3:pkg_cc_proto",
         "@envoy_api//envoy/service/discovery/v3:pkg_cc_proto",
-    ],
-)
-
-envoy_cc_extension(
-    name = "connection_handler_lib",
-    srcs = ["connection_handler_impl.cc"],
-    hdrs = [
-        "connection_handler_impl.h",
-    ],
-    # core code.
-    extra_visibility = [
-        "//test:__subpackages__",
-    ],
-    deps = [
-        "active_tcp_listener",
-        "//envoy/common:time_interface",
-        "//envoy/event:deferred_deletable",
-        "//envoy/event:dispatcher_interface",
-        "//envoy/network:client_connection_factory",
-        "//envoy/network:connection_handler_interface",
-        "//envoy/network:connection_interface",
-        "//envoy/network:filter_interface",
-        "//envoy/network:listen_socket_interface",
-        "//envoy/network:listener_interface",
-        "//envoy/runtime:runtime_interface",
-        "//envoy/server:listener_manager_interface",
-        "//envoy/stats:timespan_interface",
-        "//source/common/common:linked_object",
-        "//source/common/common:non_copyable",
-        "//source/common/event:deferred_task",
-        "//source/common/network:address_lib",
-        "//source/common/network:connection_lib",
-        "//source/common/stream_info:stream_info_lib",
-        "//source/server:active_udp_listener",
-        "//source/server:listener_manager_factory_lib",
-    ],
-)
-
-envoy_cc_library(
-    name = "active_tcp_listener",
-    srcs = ["active_tcp_listener.cc"],
-    hdrs = [
-        "active_tcp_listener.h",
-    ],
-    deps = [
-        ":active_stream_listener_base",
-        ":active_tcp_socket",
-        "//envoy/common:time_interface",
-        "//envoy/event:deferred_deletable",
-        "//envoy/event:dispatcher_interface",
-        "//envoy/network:connection_handler_interface",
-        "//envoy/network:connection_interface",
-        "//envoy/network:listen_socket_interface",
-        "//envoy/network:listener_interface",
-        "//envoy/server:listener_manager_interface",
-        "//source/common/common:assert_lib",
-        "//source/common/common:linked_object",
-        "//source/common/network:connection_lib",
-        "//source/common/stats:timespan_lib",
-        "//source/server:active_listener_base",
-    ],
-)
-
-envoy_cc_library(
-    name = "active_tcp_listener_headers",
-    hdrs = [
-        "active_stream_listener_base.h",
-        "active_tcp_listener.h",
-        "active_tcp_socket.h",
-    ],
-    deps = [
-        "//envoy/common:time_interface",
-        "//envoy/event:deferred_deletable",
-        "//envoy/event:dispatcher_interface",
-        "//envoy/event:timer_interface",
-        "//envoy/network:connection_handler_interface",
-        "//envoy/network:connection_interface",
-        "//envoy/network:filter_interface",
-        "//envoy/network:listen_socket_interface",
-        "//envoy/network:listener_interface",
-        "//envoy/server:listener_manager_interface",
-        "//source/common/common:assert_lib",
-        "//source/common/common:linked_object",
-        "//source/common/network:connection_lib",
-        "//source/common/stats:timespan_lib",
-        "//source/server:active_listener_base",
-    ],
-)
-
-envoy_cc_library(
-    name = "active_tcp_socket",
-    srcs = ["active_tcp_socket.cc"],
-    hdrs = [
-        "active_tcp_socket.h",
-    ],
-    deps = [
-        ":active_tcp_listener_headers",
-        "//envoy/common:time_interface",
-        "//envoy/event:deferred_deletable",
-        "//envoy/event:dispatcher_interface",
-        "//envoy/event:timer_interface",
-        "//envoy/network:connection_handler_interface",
-        "//envoy/network:connection_interface",
-        "//envoy/network:filter_interface",
-        "//envoy/network:listen_socket_interface",
-        "//envoy/network:listener_interface",
-        "//source/common/common:linked_object",
-        "//source/common/network:connection_lib",
-        "//source/common/network:listener_filter_buffer_lib",
-        "//source/common/stream_info:stream_info_lib",
-        "//source/server:active_listener_base",
-    ],
-)
-
-envoy_cc_library(
-    name = "active_stream_listener_base",
-    srcs = ["active_stream_listener_base.cc"],
-    hdrs = [
-        "active_stream_listener_base.h",
-    ],
-    deps = [
-        ":active_tcp_listener_headers",
-        "//envoy/common:time_interface",
-        "//envoy/event:deferred_deletable",
-        "//envoy/event:dispatcher_interface",
-        "//envoy/event:timer_interface",
-        "//envoy/network:connection_handler_interface",
-        "//envoy/network:connection_interface",
-        "//envoy/network:filter_interface",
-        "//envoy/network:listen_socket_interface",
-        "//envoy/network:listener_interface",
-        "//envoy/stream_info:stream_info_interface",
-        "//source/common/common:linked_object",
-        "//source/common/network:connection_lib",
-        "//source/common/network:listener_filter_buffer_lib",
-        "//source/server:active_listener_base",
     ],
 )
