+++ source/extensions/listener_managers/listener_manager/active_raw_udp_listener_config.cc
@@ -3,8 +3,8 @@
 #include <memory>
 #include <string>
 
-#include "source/extensions/listener_managers/listener_manager/connection_handler_impl.h"
 #include "source/server/active_udp_listener.h"
+#include "source/server/connection_handler_impl.h"
 
 namespace Envoy {
 namespace Server {
