+++ source/server/BUILD
@@ -62,6 +62,44 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
+    name = "connection_handler_lib",
+    deps = [
+        ":active_tcp_listener",
+        ":active_udp_listener",
+        ":connection_handler_impl",
+        "//envoy/network:client_connection_factory",
+        "//source/common/network:address_lib",
+    ],
+)
+
+envoy_cc_library(
+    name = "connection_handler_impl",
+    srcs = ["connection_handler_impl.cc"],
+    hdrs = [
+        "connection_handler_impl.h",
+    ],
+    deps = [
+        ":active_tcp_listener",
+        "//envoy/common:time_interface",
+        "//envoy/event:deferred_deletable",
+        "//envoy/event:dispatcher_interface",
+        "//envoy/network:connection_handler_interface",
+        "//envoy/network:connection_interface",
+        "//envoy/network:filter_interface",
+        "//envoy/network:listen_socket_interface",
+        "//envoy/network:listener_interface",
+        "//envoy/runtime:runtime_interface",
+        "//envoy/server:listener_manager_interface",
+        "//envoy/stats:timespan_interface",
+        "//source/common/common:linked_object",
+        "//source/common/common:non_copyable",
+        "//source/common/event:deferred_task",
+        "//source/common/network:connection_lib",
+        "//source/common/stream_info:stream_info_lib",
+    ],
+)
+
+envoy_cc_library(
     name = "listener_stats",
     hdrs = [
         "listener_stats.h",
@@ -86,6 +124,31 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
+    name = "active_tcp_listener",
+    srcs = ["active_tcp_listener.cc"],
+    hdrs = [
+        "active_tcp_listener.h",
+    ],
+    deps = [
+        ":active_stream_listener_base",
+        ":active_tcp_socket",
+        "//envoy/common:time_interface",
+        "//envoy/event:deferred_deletable",
+        "//envoy/event:dispatcher_interface",
+        "//envoy/network:connection_handler_interface",
+        "//envoy/network:connection_interface",
+        "//envoy/network:listen_socket_interface",
+        "//envoy/network:listener_interface",
+        "//envoy/server:listener_manager_interface",
+        "//source/common/common:assert_lib",
+        "//source/common/common:linked_object",
+        "//source/common/network:connection_lib",
+        "//source/common/stats:timespan_lib",
+        "//source/server:active_listener_base",
+    ],
+)
+
+envoy_cc_library(
     name = "active_udp_listener",
     srcs = ["active_udp_listener.cc"],
     hdrs = [
@@ -99,8 +162,83 @@ envoy_cc_library(
         "//envoy/network:listen_socket_interface",
         "//envoy/network:listener_interface",
         "//envoy/server:listener_manager_interface",
-        "//source/common/network:utility_lib",
-        "//source/server:active_listener_base",
+        "//source/server:connection_handler_impl",
+    ],
+)
+
+envoy_cc_library(
+    name = "active_tcp_listener_headers",
+    hdrs = [
+        "active_stream_listener_base.h",
+        "active_tcp_listener.h",
+        "active_tcp_socket.h",
+    ],
+    deps = [
+        ":active_listener_base",
+        "//envoy/common:time_interface",
+        "//envoy/event:deferred_deletable",
+        "//envoy/event:dispatcher_interface",
+        "//envoy/event:timer_interface",
+        "//envoy/network:connection_handler_interface",
+        "//envoy/network:connection_interface",
+        "//envoy/network:filter_interface",
+        "//envoy/network:listen_socket_interface",
+        "//envoy/network:listener_interface",
+        "//envoy/server:listener_manager_interface",
+        "//source/common/common:assert_lib",
+        "//source/common/common:linked_object",
+        "//source/common/network:connection_lib",
+        "//source/common/stats:timespan_lib",
+    ],
+)
+
+envoy_cc_library(
+    name = "active_tcp_socket",
+    srcs = ["active_tcp_socket.cc"],
+    hdrs = [
+        "active_tcp_socket.h",
+    ],
+    deps = [
+        ":active_listener_base",
+        ":active_tcp_listener_headers",
+        "//envoy/common:time_interface",
+        "//envoy/event:deferred_deletable",
+        "//envoy/event:dispatcher_interface",
+        "//envoy/event:timer_interface",
+        "//envoy/network:connection_handler_interface",
+        "//envoy/network:connection_interface",
+        "//envoy/network:filter_interface",
+        "//envoy/network:listen_socket_interface",
+        "//envoy/network:listener_interface",
+        "//source/common/common:linked_object",
+        "//source/common/network:connection_lib",
+        "//source/common/network:listener_filter_buffer_lib",
+        "//source/common/stream_info:stream_info_lib",
+    ],
+)
+
+envoy_cc_library(
+    name = "active_stream_listener_base",
+    srcs = ["active_stream_listener_base.cc"],
+    hdrs = [
+        "active_stream_listener_base.h",
+    ],
+    deps = [
+        ":active_listener_base",
+        ":active_tcp_listener_headers",
+        "//envoy/common:time_interface",
+        "//envoy/event:deferred_deletable",
+        "//envoy/event:dispatcher_interface",
+        "//envoy/event:timer_interface",
+        "//envoy/network:connection_handler_interface",
+        "//envoy/network:connection_interface",
+        "//envoy/network:filter_interface",
+        "//envoy/network:listen_socket_interface",
+        "//envoy/network:listener_interface",
+        "//envoy/stream_info:stream_info_interface",
+        "//source/common/common:linked_object",
+        "//source/common/network:connection_lib",
+        "//source/common/network:listener_filter_buffer_lib",
     ],
 )
 
@@ -370,7 +508,6 @@ envoy_cc_library(
         ":worker_lib",
         "//envoy/server:factory_context_interface",
         "//envoy/server:listener_manager_interface",
-        "//envoy/server:worker_interface",
         "//source/common/quic:quic_stat_names_lib",
     ],
 )
@@ -386,6 +523,7 @@ envoy_cc_library(
     deps = [
         ":api_listener_lib",
         ":configuration_lib",
+        ":connection_handler_lib",
         ":guarddog_lib",
         ":listener_hooks_lib",
         ":listener_manager_factory_lib",
@@ -477,8 +615,8 @@ envoy_cc_library(
     srcs = ["worker_impl.cc"],
     hdrs = ["worker_impl.h"],
     deps = [
+        ":connection_handler_lib",
         ":listener_hooks_lib",
-        ":listener_manager_factory_lib",
         "//envoy/api:api_interface",
         "//envoy/event:dispatcher_interface",
         "//envoy/event:timer_interface",
@@ -489,7 +627,6 @@ envoy_cc_library(
         "//envoy/server:worker_interface",
         "//envoy/thread:thread_interface",
         "//envoy/thread_local:thread_local_interface",
-        "//source/common/config:utility_lib",
     ],
 )
