+++ test/common/quic/BUILD
@@ -214,7 +214,6 @@ envoy_cc_test(
         "//source/common/quic:active_quic_listener_lib",
         "//source/common/quic:envoy_quic_utils_lib",
         "//source/common/quic:udp_gso_batch_writer_lib",
-        "//source/extensions/listener_managers/listener_manager:connection_handler_lib",
         "//source/extensions/quic/crypto_stream:envoy_quic_crypto_server_stream_lib",
         "//source/extensions/quic/proof_source:envoy_quic_proof_source_factory_impl_lib",
         "//source/server:configuration_lib",
@@ -241,9 +240,9 @@ envoy_cc_test(
         "//source/common/quic:envoy_quic_dispatcher_lib",
         "//source/common/quic:envoy_quic_proof_source_lib",
         "//source/common/quic:envoy_quic_server_session_lib",
-        "//source/extensions/listener_managers/listener_manager:connection_handler_lib",
         "//source/extensions/quic/crypto_stream:envoy_quic_crypto_server_stream_lib",
         "//source/server:configuration_lib",
+        "//source/server:connection_handler_lib",
         "//test/mocks/event:event_mocks",
         "//test/mocks/http:http_mocks",
         "//test/mocks/network:network_mocks",
