+++ test/extensions/bootstrap/internal_listener/BUILD
@@ -76,8 +76,8 @@ envoy_extension_cc_test(
         "//source/common/stats:stats_lib",
         "//source/extensions/bootstrap/internal_listener:active_internal_listener",
         "//source/extensions/bootstrap/internal_listener:thread_local_registry",
-        "//source/extensions/listener_managers/listener_manager:connection_handler_lib",
         "//source/extensions/transport_sockets/raw_buffer:config",
+        "//source/server:connection_handler_lib",
         "//test/mocks/access_log:access_log_mocks",
         "//test/mocks/network:network_mocks",
         "//test/server:utility_lib",
