+++ test/extensions/common/proxy_protocol/proxy_protocol_regression_test.cc
@@ -7,7 +7,7 @@
 #include "source/common/network/listen_socket_impl.h"
 #include "source/extensions/common/proxy_protocol/proxy_protocol_header.h"
 #include "source/extensions/filters/listener/proxy_protocol/proxy_protocol.h"
-#include "source/extensions/listener_managers/listener_manager/connection_handler_impl.h"
+#include "source/server/connection_handler_impl.h"
 
 #include "test/mocks/buffer/mocks.h"
 #include "test/mocks/network/mocks.h"
