+++ test/extensions/filters/listener/common/fuzz/BUILD
@@ -23,7 +23,7 @@ envoy_cc_test_library(
         ":listener_filter_fuzzer_proto_cc_proto",
         "//envoy/network:filter_interface",
         "//source/common/network:connection_balancer_lib",
-        "//source/extensions/listener_managers/listener_manager:connection_handler_lib",
+        "//source/server:connection_handler_lib",
         "//test/mocks/network:network_mocks",
         "//test/test_common:network_utility_lib",
         "//test/test_common:threadsafe_singleton_injector_lib",
