+++ test/extensions/filters/listener/common/fuzz/listener_filter_fuzzer.h
@@ -3,7 +3,7 @@
 #include "envoy/network/filter.h"
 
 #include "source/common/network/connection_balancer_impl.h"
-#include "source/extensions/listener_managers/listener_manager/connection_handler_impl.h"
+#include "source/server/connection_handler_impl.h"
 
 #include "test/extensions/filters/listener/common/fuzz/listener_filter_fakes.h"
 #include "test/extensions/filters/listener/common/fuzz/listener_filter_fuzzer.pb.validate.h"
