+++ test/extensions/filters/listener/proxy_protocol/proxy_protocol_test.cc
@@ -15,7 +15,7 @@
 #include "source/common/network/tcp_listener_impl.h"
 #include "source/common/network/utility.h"
 #include "source/extensions/filters/listener/proxy_protocol/proxy_protocol.h"
-#include "source/extensions/listener_managers/listener_manager/connection_handler_impl.h"
+#include "source/server/connection_handler_impl.h"
 
 #include "test/mocks/api/mocks.h"
 #include "test/mocks/buffer/mocks.h"
