+++ test/server/BUILD
@@ -83,7 +83,7 @@ envoy_cc_test(
         "//source/common/network:udp_packet_writer_handler_lib",
         "//source/common/stats:stats_lib",
         "//source/extensions/listener_managers/listener_manager:active_raw_udp_listener_config",
-        "//source/extensions/listener_managers/listener_manager:connection_handler_lib",
+        "//source/server:connection_handler_lib",
         "//test/mocks/access_log:access_log_mocks",
         "//test/mocks/api:api_mocks",
         "//test/mocks/network:network_mocks",
@@ -104,7 +104,7 @@ envoy_cc_test(
         "//source/common/network:address_lib",
         "//source/common/network:connection_balancer_lib",
         "//source/common/stats:stats_lib",
-        "//source/extensions/listener_managers/listener_manager:connection_handler_lib",
+        "//source/server:connection_handler_lib",
         "//test/mocks/network:io_handle_mocks",
         "//test/mocks/network:network_mocks",
         "//test/test_common:network_utility_lib",
