+++ test/server/active_tcp_listener_test.cc
@@ -8,7 +8,7 @@
 #include "source/common/network/connection_balancer_impl.h"
 #include "source/common/network/raw_buffer_socket.h"
 #include "source/common/network/utility.h"
-#include "source/extensions/listener_managers/listener_manager/active_tcp_listener.h"
+#include "source/server/active_tcp_listener.h"
 
 #include "test/mocks/common.h"
 #include "test/mocks/network/io_handle.h"
